Found a significant bug in forward loop
Hint: each neuron of a previous layer is connected to each neuron in a
next one. So input of a neuron can't be summed on demand because at a
single iteration each neuron in a previous layer adds it's signal to
input of each neuron of a next one. So final input of all neurons will
be complete only at the very and of loop.

**An item of synapses conveys connection to all neurons of a previous
layer.**